LPS-134118 Adds the filter by navigation and click event in popover to open the occurrence
This is a bit of a big change the vast majority is moving the `<StackNavigator />` states up to facilitate interactions with the Overlay and naming changes to be more consistent with what to do.

Each navigation is made a filter based on the list being shown in the current Panel is reflected in the Overlay.