LPS-165669 Fix ComponentTagTest
This test is wrong because ComponentTag does not support ESM yet, so it won't
render <script type=module> ever.

This broke when I merged a PR to only emit type=module in the presence of ESM
imports (related to the es-module-shims stuff). I wanted to be conservative and
not use type=module unless needed.

I'm not sure why this test hasn't failed since that PR, but I guess it's because
the CI decided not to run it.