LPD-46900 Remove usage of LayoutRevisionLocalService.fetchLayoutRevision(layoutSetBranchId, layoutBranchId, head, plid)
This usage of the LayoutRevision.L_L_H_P does not make sense. The finder returns
a single object but does not use a unique index, meaning the finder's output is
can produce multiple results (it is only logically unique). Instead, it should
use the already fetched list of layout revisions and filter by the
layoutBranchId. This change should produce the same output as the list is sorted
the same way as the finder output is.