LPD-60643 Semantic versioning
# breaking

## What portal-kernel/src/com/liferay/portal/kernel/model/LayoutSet.java

The return type of getVirtualHostnames() is changed from TreeMap to NavigableMap. The parameter type of setVirtualHostnames() is changed from TreeMap to NavigableMap.

## Why

To hide the internal details in order to avoid unncessary map copying.

----

# breaking

## What portal-kernel/src/com/liferay/portal/kernel/util/Portal.java

The return type of getVirtualHostnames() is changed from TreeMap to NavigableMap.

## Why

To adapt to the Layout api change.

----